Update talk date test
Update the test for updating a talk node as it was essentially doing the
same as the test for creating a new talk node.

The test now ensures that the talk is saved before adding an event, and
that the talk created date is different to the original created date as
well as that it matches the latest event date.