Fix Plausible loading for authenticated users
Return early within the Plausible if the user is authenticated so that
it doesn't pick up when I'm adding or editing content on the site.

Though the role is configured in the Plausible module's admin form, the
conditions in the code don't exclude it. I'll look into this further and
submit a patch upstream along with some tests.

References #350